style(cluster): move ca-cert into manager_util.go
