+++ pkg/cluster/manager.go
@@ -16,8 +16,6 @@ package cluster
 import (
 	"bytes"
 	"crypto/tls"
-	"crypto/x509"
-	"encoding/pem"
 	"errors"
 	"fmt"
 	"io"
@@ -43,7 +41,6 @@ import (
 	"github.com/pingcap/tiup/pkg/crypto"
 	"github.com/pingcap/tiup/pkg/environment"
 	"github.com/pingcap/tiup/pkg/errutil"
-	"github.com/pingcap/tiup/pkg/file"
 	"github.com/pingcap/tiup/pkg/logger/log"
 	"github.com/pingcap/tiup/pkg/meta"
 	pkgver "github.com/pingcap/tiup/pkg/repository/version"
@@ -2430,81 +2427,4 @@ func refreshMonitoredConfigTask(
 		}
 	}
 	return tasks
-}
-
-func genAndSaveClusterCA(clusterName, tlsPath string) (*crypto.CertificateAuthority, error) {
-	ca, err := crypto.NewCA(clusterName)
-	if err != nil {
-		return nil, err
-	}
-
-	// save CA private key
-	if err := file.SaveFileWithBackup(filepath.Join(tlsPath, spec.TLSCAKey), ca.Key.Pem(), ""); err != nil {
-		return nil, perrs.Annotatef(err, "cannot save CA private key for %s", clusterName)
-	}
-
-	// save CA certificate
-	if err := file.SaveFileWithBackup(
-		filepath.Join(tlsPath, spec.TLSCACert),
-		pem.EncodeToMemory(&pem.Block{
-			Type:  "CERTIFICATE",
-			Bytes: ca.Cert.Raw,
-		}), ""); err != nil {
-		return nil, perrs.Annotatef(err, "cannot save CA certificate for %s", clusterName)
-	}
-
-	return ca, nil
-}
-
-func genAndSaveClientCert(ca *crypto.CertificateAuthority, clusterName, tlsPath string) error {
-	privKey, err := crypto.NewKeyPair(crypto.KeyTypeRSA, crypto.KeySchemeRSASSAPSSSHA256)
-	if err != nil {
-		return perrs.AddStack(err)
-	}
-
-	// save client private key
-	if err := file.SaveFileWithBackup(filepath.Join(tlsPath, spec.TLSClientKey), privKey.Pem(), ""); err != nil {
-		return perrs.Annotatef(err, "cannot save client private key for %s", clusterName)
-	}
-
-	csr, err := privKey.CSR(
-		"tiup-cluster-client",
-		fmt.Sprintf("%s-client", clusterName),
-		[]string{}, []string{},
-	)
-	if err != nil {
-		return perrs.Annotatef(err, "cannot generate CSR of client certificate for %s", clusterName)
-	}
-	cert, err := ca.Sign(csr)
-	if err != nil {
-		return perrs.Annotatef(err, "cannot sign client certificate for %s", clusterName)
-	}
-
-	// save client certificate
-	if err := file.SaveFileWithBackup(
-		filepath.Join(tlsPath, spec.TLSClientCert),
-		pem.EncodeToMemory(&pem.Block{
-			Type:  "CERTIFICATE",
-			Bytes: cert,
-		}), ""); err != nil {
-		return perrs.Annotatef(err, "cannot save client PEM certificate for %s", clusterName)
-	}
-
-	// save pfx format certificate
-	clientCert, err := x509.ParseCertificate(cert)
-	if err != nil {
-		return perrs.Annotatef(err, "cannot decode signed client certificate for %s", clusterName)
-	}
-	pfxData, err := privKey.PKCS12(clientCert, ca)
-	if err != nil {
-		return perrs.Annotatef(err, "cannot encode client certificate to PKCS#12 format for %s", clusterName)
-	}
-	if err := file.SaveFileWithBackup(
-		filepath.Join(tlsPath, spec.PFXClientCert),
-		pfxData,
-		""); err != nil {
-		return perrs.Annotatef(err, "cannot save client PKCS#12 certificate for %s", clusterName)
-	}
-
-	return nil
 }
